Refactor SignatureRequestMessage for queue
